Add automounting of the serviceaccount token
This is an optional boolean, on newer versions of terraform and the
kubernetes provider this default to `true`. However older versions don't
have this set to true and therefore default to `false`

This ultimately result in deployments failing due to them not being able
to find the token.

By setting it to `true` we can ensure it's mounted no matter what
version of terraform and kubernetes provider is used.