Remove conditions from policy statements regarding tags
This patch temporarily removes these conditions because they prevent the
load balancer controller from working properly. The controller tries to
add tags and this fails. The controller than no longer updates the
target group registrations, which means the traffic can not be
forwarded.

If this issue is resolved, the code should be updated accordingly.